Epic

+++ Point.java
@@ -9,8 +9,8 @@ public class Point {
         this.y = y;
     }
 
-    public int calculateDistance(Point point2) {
-        return Math.sqrt(Math.pow(x + point2.getX, 2) + Math.pow(y + point2.getY, 2));
+    public double calculateDistance(Point point2) {
+        return Math.sqrt(Math.pow(x + point2.getX(), 2) + Math.pow(y + point2.getY(), 2));
     }
 
     public int getX() {
